Fixed broken tests after rebase
