+++ Sources/TuistCoreTesting/Utils/MockBinaryLocator.swift
@@ -16,4 +16,13 @@ public final class MockBinaryLocator: BinaryLocating {
         }
         return stubbedSwiftLintPathResult
     }
+    
+    public var swiftDocPathStub: (() throws -> AbsolutePath)?
+    public func swiftDocPath() throws -> AbsolutePath {
+        if let swiftDocPathStub = swiftDocPathStub {
+            return try swiftDocPathStub()
+        } else {
+            throw BinaryLocatorError.swiftDocNotFound
+        }
+    }
 }
